requirements.txt
openai
yfinance
requests
pandas
beautifulsoup4
python-dotenv
streamlit

